+++ src/config.rs
@@ -91,6 +91,7 @@ pub enum Protocol {
     Postgres,
     Mysql,
     Sqlserver,
+    Teradata,
 }
 
 impl Config {
@@ -99,6 +100,7 @@ impl Config {
             Protocol::Postgres => 5432,
             Protocol::Mysql => 3306,
             Protocol::Sqlserver => 1433,
+            Protocol::Teradata => 1025,  // Default Teradata port
         })
     }